modifying run.py

+++ run.py
@@ -6,14 +6,17 @@ from bson.objectid import ObjectId
 import boto3
 
 app = Flask(__name__)
-app.config["MONGO_URI"] = "mongodb+srv://daradona10:maradona1986@myfirstcluster-agrgt.mongodb.net/Milestone-Project3?retryWrites=true&w=majority"
+app.config[
+    "MONGO_URI"] = "mongodb+srv://daradona10:maradona1986@myfirstcluster-agrgt.mongodb.net/Milestone-Project3?retryWrites=true&w=majority"
 mongo = PyMongo(app)
 
 UPLOAD_FOLDER = "recipe_images"
 
+
 @app.route("/")
 def get_meals():
     return render_template("meals.html")
+
 
 @app.route("/get_mains")
 def get_mains():
@@ -21,8 +24,10 @@ def get_mains():
     course_types = get_course_types()
     return render_template("maincourses.html", dishes=starters, course_types=course_types)
 
+
 def get_course_types():
     return ['starters', 'main courses', 'sides', 'desserts']
+
 
 def get_recipes_by_type(recipe_type):
     recipes_in_db = mongo.db.recipes.find({"course_type": recipe_type})
@@ -38,10 +43,12 @@ def get_recipes_by_type(recipe_type):
             "ingredients": s["ingredients"],
             "method": s["method"],
             "course_type": s["course_type"],
-            "image_path": "https://daradona-milestone-project3.s3.amazonaws.com/" + s["image_path"] if "image_path" in s else ""
+            "image_path": "https://daradona-milestone-project3.s3.amazonaws.com/" + s[
+                "image_path"] if "image_path" in s else ""
         })
-    
+
     return recipes
+
 
 @app.route("/get_starters")
 def get_starters():
@@ -49,11 +56,13 @@ def get_starters():
     course_types = get_course_types()
     return render_template("starters.html", dishes=starters, course_types=course_types)
 
+
 @app.route("/get_desserts")
 def get_desserts():
     desserts = get_recipes_by_type('desserts')
     course_types = get_course_types()
-    return render_template("desserts.html", dishes=desserts, course_types=course_types)  
+    return render_template("desserts.html", dishes=desserts, course_types=course_types)
+
 
 @app.route("/get_sides")
 def get_sides():
@@ -61,23 +70,26 @@ def get_sides():
     course_types = get_course_types()
     return render_template("sides.html", dishes=sides, course_types=course_types)
 
+
 @app.route("/add_meal")
 def add_meal():
     course_types = get_course_types()
     return render_template("add_meal.html", course_types=course_types)
 
+
 def upload_file_to_s3(filename):
     print('Uploading file to s3')
     try:
         s3_client = boto3.client('s3',
-            aws_access_key_id='AKIASIXFUF7B5ETRWRH5',
-            aws_secret_access_key='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',
-        )
-        s3_client.upload_file(filename, 'daradona-milestone-project3', filename, ExtraArgs={'ACL':'public-read'})
+                                 aws_access_key_id='AKIASIXFUF7B5ETRWRH5',
+                                 aws_secret_access_key='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',
+                                 )
+        s3_client.upload_file(filename, 'daradona-milestone-project3', filename, ExtraArgs={'ACL': 'public-read'})
     except Exception as e:
         print(e)
 
     print("Uploaded file successfully!")
+
 
 @app.route("/add_new_recipe", methods=['GET', 'POST'])
 def add_new_recipe():
@@ -91,14 +103,14 @@ def add_new_recipe():
     course_type = request.form.get('course_type')
 
     f = request.files['file_name']
-    file_path = os.path.join(UPLOAD_FOLDER, f.filename)
-    f.save(file_path)
-    upload_file_to_s3("recipe_images/" + f.filename)
+    if f:
+        file_path = os.path.join(UPLOAD_FOLDER, f.filename)
+        f.save(file_path)
+        upload_file_to_s3("recipe_images/" + f.filename)
+        os.remove(file_path)
+        print('Removed local file successfully.')
 
-    os.remove(file_path)
-    print('Removed local file successfully.')
-    
-    mongo.db.recipes.insert_one({
+    new_values = {
         'name': name,
         'description': description,
         'prep_time': prep_time,
@@ -106,12 +118,17 @@ def add_new_recipe():
         'makes': makes,
         'ingredients': ingredients,
         'method': method,
-        'course_type': course_type,
-        'image_path': "recipe_images/" + f.filename
-    })
+        'course_type': course_type
+    }
+
+    if f:
+        new_values['image_path'] = "recipe_images/" + f.filename
+
+    mongo.db.recipes.insert_one(new_values)
 
     f.close()
     return redirect(url_for('add_meal'))
+
 
 @app.route('/delete_recipe/<recipe_id>')
 def delete_recipe(recipe_id):
@@ -132,6 +149,7 @@ def delete_recipe(recipe_id):
 
     return redirect(url_for(template))
 
+
 @app.route("/update_recipe", methods=['PUT'])
 def update_recipe():
     json_body = request.get_json()
@@ -139,13 +157,13 @@ def update_recipe():
     dish_type = json_body["dish_type"]
     dish = json_body["dish"]
 
-    myquery = { '_id': ObjectId(dish_id) }
-    newvalues = { "$set": dish }
+    myquery = {'_id': ObjectId(dish_id)}
+    newvalues = {"$set": dish}
     mongo.db.recipes.update_one(myquery, newvalues)
 
     dishes = get_recipes_by_type(dish_type)
     course_types = get_course_types()
-    return jsonify(dishes = dishes, course_types=course_types)
+    return jsonify(dishes=dishes, course_types=course_types)
 
 
 @app.route('/edit_recipe/<recipe_id>')
@@ -161,16 +179,16 @@ def edit_single_recipe(recipe_id):
     course_type = request.form.get('course_type')
 
     f = request.files['file_name']
-    file_path = os.path.join(UPLOAD_FOLDER, f.filename)
-    f.save(file_path)
-    upload_file_to_s3("recipe_images/" + f.filename)
-
-    os.remove(file_path)
-    print('Removed local file successfully.')
-    f.close()
+    if f:
+        file_path = os.path.join(UPLOAD_FOLDER, f.filename)
+        f.save(file_path)
+        upload_file_to_s3("recipe_images/" + f.filename)
+        os.remove(file_path)
+        print('Removed local file successfully.')
+        f.close()
 
     myquery = {'_id': ObjectId(recipe_id)}
-    newvalues = {"$set": {
+    updated_values = {
         "dish_id": recipe_id,
         "course_type": course_type,
         "name": request.form.get('recipe_name'),
@@ -179,9 +197,13 @@ def edit_single_recipe(recipe_id):
         "cooking_time": request.form.get('cooking_time'),
         "makes": request.form.get('makes'),
         "ingredients": request.form.get('ingredients'),
-        "method": request.form.get('method'),
-        "image_path": "recipe_images/" + f.filename
-    }}
+        "method": request.form.get('method')
+    }
+
+    if f:
+        updated_values["image_path"] = "recipe_images/" + f.filename
+
+    newvalues = {"$set": updated_values}
     mongo.db.recipes.update_one(myquery, newvalues)
 
     template = 'get_sides'
@@ -193,6 +215,7 @@ def edit_single_recipe(recipe_id):
         template = 'get_desserts'
 
     return redirect(url_for(template))
+
 
 if __name__ == "__main__":
     app.run(host=os.environ.get("IP"), port=8000, debug=True)
